Add WM.deleteInSingleQuote() & purge RcR's seek fns
While implementing RcR's seekBackwards(), my tests didn't account for
single character leafs.
Even if it did, when it move backwards to a leaf of a previous line, it
doesn't know the number of characters on that line, which makes it
useless.
The only way to know a line's noc is to walk from line beginning to its
end and sum all the noc. Which defeat the purpose of
walkLineColBackwards() entirely.

After purging RcR's seekForward() & seekBackwards(), I implement them
inside RopeMan instead.
Although I have to allocate memory, this way is much simpler.
I wish I didn't have to allocate memory, which is totally possible if I
write a method in RcR that walk from line begin and check now I think
about it.
==> Still walk from line begin, but no need to allocate.

I'll revisit this topic later. At the moment, I'll settle for the less
optimal implementation in RopeMan.