chore: add 'Meslo LG L DZ Regular' font for demo
Due to the font being hard coded into `main.zig`, in order to make the
demo work, I add 'Meslo LG L DZ Regular' font file to source control.

This is oviously not ideal. I will try to find solutions for users to
load their own font files. In the mean time, this is a bandaid for the
demo to work.

No copyright infringement intended. If you have legal rights to the font
and want me to take it down, I will.

Fix Discord link